SonarQube Cloud analysis, native code documentation
This documents on how to run a SonarQube Cloud analysis, analyzing the native
code for macOS only for now. Also more documentation on how to build both on
macOS and Windows the respective project.

The Eclipse CDT project now also provides a post-build step to copy the shared
library into the resources folder.